[FEAT] Add Extension Folder
