+++ Bubblia/Bubblia/HandGestureProcessor.swift
@@ -62,14 +62,3 @@ class HandGestureProcessor {
 }
 
 // MARK: - CGPoint helpers
-
-extension CGPoint {
-
-    static func midPoint(p1: CGPoint, p2: CGPoint) -> CGPoint {
-        return CGPoint(x: (p1.x + p2.x) / 2, y: (p1.y + p2.y) / 2)
-    }
-    
-    func distance(from point: CGPoint) -> CGFloat {
-        return hypot(point.x - x, point.y - y)
-    }
-}
